+++ semcore/d3-chart/src/component/Chart/DonutChart.type.ts
@@ -8,11 +8,17 @@ import type { BaseChartProps } from './AbstractChart.type';
 export type DonutChartData = Record<string, number>;
 
 export type DonutChartProps = BaseChartProps<DonutChartData> & {
+  /** Internal */
   groupKey?: never;
+  /** Custom x-axis scale */
   xScale?: ScaleLinear<any, any>;
+  /** Custom y-axis scale */
   yScale?: ScaleLinear<any, any>;
+  /** Controls the inner radius of the donut */
   innerRadius?: number;
+  /** Creates a semi-donut chart when enabled */
   halfsize?: boolean;
+  /** Content displayed in the center of the donut */
   innerLabel?: React.ReactNode;
 };